fix mistake

+++ BenchmarkTest.cpp
@@ -29,7 +29,7 @@ void BenchmarkTest::pi_device(double& result) {
 
 void BenchmarkTest::pi_host(double& result) {
     int count = 0;
-    for (float i = 0; i < samples; i++) {
+    for (int i = 0; i < samples; i++) {
         float x = float(rand()) / float(RAND_MAX);
         float y = float(rand()) / float(RAND_MAX);
         if (sqrt(x * x + y * y) < 1) count++;
